refractor create export-config component
