+++ apps/react-app/src/common-app/components/export-config/index.ts
@@ -1,0 +1,1 @@
+export * from './export-config.component';
