+++ apps/react-app/src/common-app/components/index.ts
@@ -4,3 +4,4 @@ export * from './button';
 export * from './navbar';
 export * from './card';
 export * from './modal';
+export * from './export-config';
